BaseCollection new id (#21)
* add IdVersion to XWikiDocument and BaseCollection

* move IdVersion to celements package

* fix hbm

* fix ConfigurationSource in NoStackTracePatternLayout

* move hbm's to celements-xwiki-core and remove mssql

* move hbm's to celements-xwiki-core and remove mssql

* fix ConfigurationSource in NoStackTracePatternLayout

* extract verifyIdVersion

* change object id to long

* move idVersion to BaseElement

* move id to BaseElement

* add BaseElement hasValidId

* add computeNextObjectId

* add TODO comment to XWikiDocument.idVersion

* add XWikiDocument#hasValidId

* change HibernateStore from int to long

* use IdComputer in CelHibernateStore

* extract 'collectVersionedObjectIds'

* change naming in tests

* remove unnecessary import java.util.Optional

* update parent dependency

* remove unused Optional import

* add manifest

* log ObjectNotFoundException

* add BaseCollectionIdToLongMigration

* add BaseCollectionIdToLongMigrationTest

* rename migration

* remove newline

* move ConcurrentCacheTest from celements-core

* fix ConcurrentCacheTest

* improve logging

* load property with id

* fix hibernate access for properties

* improve logging in CelHibernateStore

* BaseCollectionIdMigration drop/add FKs (#18)

* refactor executeWriteSql transaction and exception handling

* add executeReadSql

* BaseCollectionIdMigration drop/add FK

* fix Joiner import

* catching exception for stack trace

* extract migrateTable

* fix executeReadSql

* improve foreign key readd

* increase logging for loadForeignKeys

* add getDatabaseWithPrefix

* add recursive table migration

* improve logging

* rename add/drop methods

* add test test_migrate_fail

* improve logging

* BaseCollectionIdColumnMigration mapped tables (#19)

* migration for mapped classes

* rename migration

* improve logging

* change logging

* change logging

* extract InformationSchema

* change types to long

* improv logging levels

* log invalid id objects

* fix logging

* clone BaseElement id

* improve logXWikiDoc

* fix Iterables.concat

* change logging

* change logging

* change logging

* change logging

* fix BaseObject clone and duplicate

* CelementsIdComputer change computeId behaviour (#20)

* change computeId behaviour to 50bit | 2bit | 12bit

* add computeMaxDocumentId

* add fast failure for possible XWIKI_2 ids

* extract verifyId

* update to celements 3

* update to celements 3

* update to celements 3

* add dependency xwiki-core-observation-api

+++ celements-model/src/test/java/com/celements/store/TestHibernateQuery.java
@@ -132,7 +132,7 @@ public abstract class TestHibernateQuery<T> extends AbstractQueryImpl {
   }
 
   public static void expectLoadProperties(Session sessionMock, final List<BaseObject> objList,
-      final Map<Integer, List<String[]>> propertiesMap) {
+      final Map<Long, List<String[]>> propertiesMap) {
     String hql = "select prop.name, prop.classType from BaseProperty as prop where prop.id.id = :id";
     Query queryProp = new TestHibernateQuery<String[]>(hql) {
 
@@ -149,7 +149,7 @@ public abstract class TestHibernateQuery<T> extends AbstractQueryImpl {
       @Override
       public Object answer() throws Throwable {
         BaseProperty property = (BaseProperty) getCurrentArguments()[0];
-        Integer objId = property.getObject().getId();
+        Long objId = property.getObject().getId();
         for (BaseObject templBaseObject : objList) {
           if (objId.equals(templBaseObject.getId())) {
             for (Object theObj : templBaseObject.getFieldList()) {
